+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Summary
 #### Features
 
+- Define privacy flags for the daemon with the help of the `sendmail::privacy_flags` class.
 - Add parameter `daemon_name` to the `sendmail::mc::daemon_options` type. This allows reusing the same name for multiple daemon option enties.
 - Allow a string parameter for `sendmail::mc::feature` if the feature requires only a single argument.